+++ feature/main/src/main/java/kr/co/main/calendar/providers/FakeMainCalendarDataProvider.kt
@@ -4,7 +4,7 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import kr.co.domain.entity.HolidayEntity
-import kr.co.main.calendar.calendar.maincalendar.MainCalendarData
+import kr.co.main.calendar.ui.common.maincalendar.MainCalendarData
 import kr.co.main.calendar.model.CropModel
 import kr.co.main.calendar.model.DiaryModel
 import kr.co.main.calendar.model.ScheduleModel
