+++ feature/main/src/main/java/kr/co/main/calendar/ui/diarytab/adddiary/diaryinput/DiaryWorkInput.kt
@@ -1,4 +1,4 @@
-package kr.co.main.calendar.adddiary.diaryinput
+package kr.co.main.calendar.ui.diarytab.adddiary.diaryinput
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
@@ -38,7 +38,7 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import kr.co.main.R
-import kr.co.main.calendar.common.CalendarUnderLineTextField
+import kr.co.main.calendar.ui.common.CalendarUnderLineTextField
 import kr.co.main.calendar.model.DiaryModel
 import kr.co.main.calendar.providers.FakeWorkDescriptionModelListProvider
 import kr.co.ui.icon.DreamIcon
